# docs/swagger.yaml
definitions:
...
  models.PostCoverResponse:
    description: Response model for post cover upload
    properties:
      cover:
        example: https://example.com/cover.jpg
        type: string
    type: object
...
  models.StandardResponse:
    description: A standard API response format with data payload
    properties:
      data: {}
      error:
        example: Invalid input
        type: string
      message:
        example: Operation completed successfully
        type: string
      status:
        example: success
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /posts/{id}/cover:
    delete:
      description: Remove the cover image from a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cover deleted successfully
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Post not found
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Delete post cover image
      tags:
      - Posts
    post:
      consumes:
      - multipart/form-data
      description: Upload a new cover image for a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cover image file (JPG, JPEG, PNG, WEBP, max 5MB)
        in: formData
        name: cover
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Cover uploaded successfully
          schema:
            $ref: '#/definitions/models.PostCoverResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Post not found
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Upload post cover image
      tags:
      - Posts
...
  /tags:
    get:
      description: Returns all tags with their post counts
      produces:
      - application/json
      responses:
        "200":
          description: List of tags with post counts
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Get all tags
      tags:
      - Tags
  /tags/popular:
    get:
      description: Returns the most used tags with post counts (limited to 10)
      produces:
      - application/json
      responses:
        "200":
          description: List of popular tags with post counts
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Get popular tags
      tags:
      - Tags
swagger: "2.0"